+++ src/controllers/auth.js
@@ -1,9 +1,8 @@
-const generate = require('nanoid/generate');
-const slugify = require('slugify');
+const { customAlphabet } = require('nanoid/async');
 const moment = require('moment');
 const sendgrid = require('../config/sendgrid');
 
-const alphabet =
+const urlFriendyAlphabet =
   '0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz';
 
 /**
@@ -34,11 +33,11 @@ exports.postSignup = async (req, res) => {
     });
 
     // Set the token and the expire date.
-    const token = await generate(alphabet, 24);
-    const tokenExpire = moment().add('3', 'h');
+    const token = customAlphabet(urlFriendyAlphabet, 32);
+    const tokenExpire = moment().add('1', 'h');
 
     // Sets the token and expire date to the database
-    user.emailVerificationToken = token;
+    user.emailVerificationToken = await token();
     user.emailVerificationTokenExpire = tokenExpire;
     await user.save();