+++ src/controllers/index.js
@@ -1,4 +1,3 @@
-/* eslint-disable indent */
 const path = require('path');
 const fs = require('fs-extra');