+++ src/emails/AdminCreateUserActivation.js
@@ -2,9 +2,10 @@ const mjml = require('mjml');
 
 /**
  *  @param token
+ *  @param password
  *  Express is the domain from the request.
  */
-module.exports = token =>
+module.exports = (token, password) =>
   mjml(`<mjml>
   <mj-head>
        <mj-attributes>
@@ -33,6 +34,7 @@ module.exports = token =>
               <mj-column>
                 <mj-text>Please click activate to finalize your account creation.</mj-text>
                 <mj-text>If you did not request this account to be made or want your data removed. Please click the delete button.</mj-text>
+                <mj-text>Temp password for your account is: ${password}</mj-text>
               </mj-column>
             </mj-section>
             <mj-section>