+++ src/validation/admin/userUpdate.js
@@ -2,7 +2,6 @@ const Validator = require('validator');
 const isEmpty = require('../isEmpty');
 
 module.exports = (req, res, next) => {
-  // eslint-disable-next-line object-curly-newline
   let { username, email, newPassword } = req.body;
 
   // eslint-disable-next-line prefer-const