+++ src/validation/api/v1/delete.js
@@ -1,6 +1,7 @@
 const isEmpty = require('../../isEmpty');
 
 module.exports = (req, res, next) => {
+
   // eslint-disable-next-line prefer-const
   let errors = {};